feat(anvil): support integer block numbers
