+++ anvil/core/src/eth/mod.rs
@@ -11,16 +11,19 @@ use ethers_core::{
     abi::ethereum_types::H64,
     types::{Address, BlockId, BlockNumber, Bytes, TxHash, H256, U256},
 };
-use serde::{Deserialize, Deserializer};
+use serde::Deserialize;
+use serde_helpers::Params;
 
 pub mod block;
 pub mod call;
 pub mod filter;
 pub mod receipt;
+mod serde_helpers;
 pub mod subscription;
 pub mod transaction;
 pub mod trie;
 pub mod utils;
+use serde_helpers::*;
 
 /// Represents ethereum JSON-RPC API
 #[derive(Clone, Debug, PartialEq, Deserialize)]
@@ -60,7 +63,7 @@ pub enum EthRequest {
     EthGetBlockByHash(H256, bool),
 
     #[serde(rename = "eth_getBlockByNumber")]
-    EthGetBlockByNumber(BlockNumber, bool),
+    EthGetBlockByNumber(#[serde(deserialize_with = "lenient_block_number")] BlockNumber, bool),
 
     #[serde(rename = "eth_getTransactionCount")]
     EthGetTransactionCount(Address, Option<BlockId>),
@@ -68,13 +71,19 @@ pub enum EthRequest {
     #[serde(rename = "eth_getBlockTransactionCountByHash")]
     EthGetTransactionCountByHash(H256),
 
-    #[serde(rename = "eth_getBlockTransactionCountByNumber")]
+    #[serde(
+        rename = "eth_getBlockTransactionCountByNumber",
+        deserialize_with = "lenient_block_number_seq"
+    )]
     EthGetTransactionCountByNumber(BlockNumber),
 
     #[serde(rename = "eth_getUncleCountByBlockHash")]
     EthGetUnclesCountByHash(H256),
 
-    #[serde(rename = "eth_getUncleCountByBlockNumber")]
+    #[serde(
+        rename = "eth_getUncleCountByBlockNumber",
+        deserialize_with = "lenient_block_number_seq"
+    )]
     EthGetUnclesCountByNumber(BlockNumber),
 
     #[serde(rename = "eth_getCode")]
@@ -106,7 +115,10 @@ pub enum EthRequest {
     EthGetTransactionByBlockHashAndIndex(TxHash, Index),
 
     #[serde(rename = "eth_getTransactionByBlockNumberAndIndex")]
-    EthGetTransactionByBlockNumberAndIndex(BlockNumber, Index),
+    EthGetTransactionByBlockNumberAndIndex(
+        #[serde(deserialize_with = "lenient_block_number")] BlockNumber,
+        Index,
+    ),
 
     #[serde(rename = "eth_getTransactionReceipt", with = "sequence")]
     EthGetTransactionReceipt(H256),
@@ -115,7 +127,10 @@ pub enum EthRequest {
     EthGetUncleByBlockHashAndIndex(H256, Index),
 
     #[serde(rename = "eth_getUncleByBlockNumberAndIndex")]
-    EthGetUncleByBlockNumberAndIndex(BlockNumber, Index),
+    EthGetUncleByBlockNumberAndIndex(
+        #[serde(deserialize_with = "lenient_block_number")] BlockNumber,
+        Index,
+    ),
 
     #[serde(rename = "eth_getLogs", with = "sequence")]
     EthGetLogs(Filter),
@@ -171,7 +186,7 @@ pub enum EthRequest {
     TraceTransaction(H256),
 
     /// Trace transaction endpoint for parity's `trace_block`
-    #[serde(rename = "trace_block", with = "sequence")]
+    #[serde(rename = "trace_block", deserialize_with = "lenient_block_number_seq")]
     TraceBlock(BlockNumber),
 
     // Custom endpoints, they're not extracted to a separate type out of serde convenience
@@ -375,121 +390,6 @@ pub enum EthPubSub {
 pub enum EthRpcCall {
     Request(Box<EthRequest>),
     PubSub(EthPubSub),
-}
-
-#[derive(Deserialize)]
-#[serde(untagged)]
-enum Numeric {
-    U256(U256),
-    Num(u64),
-}
-
-impl From<Numeric> for U256 {
-    fn from(n: Numeric) -> U256 {
-        match n {
-            Numeric::U256(n) => n,
-            Numeric::Num(n) => U256::from(n),
-        }
-    }
-}
-
-#[derive(Deserialize)]
-#[serde(untagged)]
-enum NumericSeq {
-    Seq([Numeric; 1]),
-    U256(U256),
-    Num(u64),
-}
-
-/// Deserializes single integer params: `1, [1], ["0x01"]`
-fn deserialize_number_seq<'de, D>(deserializer: D) -> Result<U256, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let num = match NumericSeq::deserialize(deserializer)? {
-        NumericSeq::Seq(seq) => seq.into_iter().next().unwrap().into(),
-        NumericSeq::U256(n) => n,
-        NumericSeq::Num(n) => U256::from(n),
-    };
-
-    Ok(num)
-}
-
-fn deserialize_number<'de, D>(deserializer: D) -> Result<U256, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    Numeric::deserialize(deserializer).map(Into::into)
-}
-
-fn deserialize_number_opt<'de, D>(deserializer: D) -> Result<Option<U256>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let num = match Option::<Numeric>::deserialize(deserializer)? {
-        Some(Numeric::U256(n)) => Some(n),
-        Some(Numeric::Num(n)) => Some(U256::from(n)),
-        _ => None,
-    };
-
-    Ok(num)
-}
-
-#[derive(Debug, Clone, PartialEq, Deserialize)]
-pub struct Params<T> {
-    pub params: T,
-}
-
-#[allow(unused)]
-mod sequence {
-    use serde::{
-        de::DeserializeOwned, ser::SerializeSeq, Deserialize, Deserializer, Serialize, Serializer,
-    };
-
-    #[allow(unused)]
-    pub fn serialize<S, T>(val: &T, s: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-        T: Serialize,
-    {
-        let mut seq = s.serialize_seq(Some(1))?;
-        seq.serialize_element(val)?;
-        seq.end()
-    }
-
-    pub fn deserialize<'de, T, D>(d: D) -> Result<T, D::Error>
-    where
-        D: Deserializer<'de>,
-        T: DeserializeOwned,
-    {
-        let mut seq = Vec::<T>::deserialize(d)?;
-        if seq.len() != 1 {
-            return Err(serde::de::Error::custom(format!(
-                "expected params sequence with length 1 but got {}",
-                seq.len()
-            )))
-        }
-        Ok(seq.remove(0))
-    }
-}
-
-/// A module that deserializes `[]` optionally
-mod empty_params {
-    use serde::{Deserialize, Deserializer};
-
-    pub fn deserialize<'de, D>(d: D) -> Result<(), D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let seq = Option::<Vec<()>>::deserialize(d)?.unwrap_or_default();
-        if !seq.is_empty() {
-            return Err(serde::de::Error::custom(format!(
-                "expected params sequence with length 0 but got {}",
-                seq.len()
-            )))
-        }
-        Ok(())
-    }
 }
 
 #[cfg(test)]
@@ -975,6 +875,27 @@ mod tests {
         let s = r#"{"method": "eth_getBalance", "params": ["0x295a70b2de5e3953354a6a8344e616ed314d7251", "latest"]}"#;
         let value: serde_json::Value = serde_json::from_str(s).unwrap();
 
+        let _req = serde_json::from_value::<EthRequest>(value).unwrap();
+    }
+
+    #[test]
+    fn test_serde_eth_block_by_number() {
+        let s = r#"{"method": "eth_getBlockByNumber", "params": ["0x0", true]}"#;
+        let value: serde_json::Value = serde_json::from_str(s).unwrap();
+        let _req = serde_json::from_value::<EthRequest>(value).unwrap();
+        let s = r#"{"method": "eth_getBlockByNumber", "params": ["latest", true]}"#;
+        let value: serde_json::Value = serde_json::from_str(s).unwrap();
+        let _req = serde_json::from_value::<EthRequest>(value).unwrap();
+        let s = r#"{"method": "eth_getBlockByNumber", "params": ["earliest", true]}"#;
+        let value: serde_json::Value = serde_json::from_str(s).unwrap();
+        let _req = serde_json::from_value::<EthRequest>(value).unwrap();
+        let s = r#"{"method": "eth_getBlockByNumber", "params": ["pending", true]}"#;
+        let value: serde_json::Value = serde_json::from_str(s).unwrap();
+        let _req = serde_json::from_value::<EthRequest>(value).unwrap();
+
+        // this case deviates from the spec, but we're supporting this for legacy reasons: <https://github.com/foundry-rs/foundry/issues/1868>
+        let s = r#"{"method": "eth_getBlockByNumber", "params": [0, true]}"#;
+        let value: serde_json::Value = serde_json::from_str(s).unwrap();
         let _req = serde_json::from_value::<EthRequest>(value).unwrap();
     }
 }
